requirements-dev.txt
-r requirements.txt
pytest~=6.1.1
pytest-django~=3.10.0
pytest-dotenv~=0.5.2
